requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
fastapi==0.100.0
h11==0.14.0
harperdb==0.4.2
idna==3.4
pydantic==2.0.3
pydantic_core==2.3.0
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.1
